Oauth with tests implemented
