+++ app.js
@@ -7,6 +7,8 @@ var path = require('path')
 
 var app = express()
 
+var auth = require('./services/auth')
+
 // initialize swagger-jsdoc
 var swaggerSpec = require('./swagger/swagger.js')
 
@@ -16,18 +18,14 @@ app.get('/swagger.json', function (req, res) {
   res.send(swaggerSpec)
 })
 
+app.options('*', cors())
 app.use(cors())
+
 app.use(bodyParser.urlencoded({
   extended: true
 }))
 
 app.use(bodyParser.json())
-
-app.use(function (req, res, next) {
-  res.header('Access-Control-Allow-Origin', '*')
-  res.header('Access-Control-Allow-Headers', 'Origin, X-Requested-With, Content-Type, Accept')
-  next()
-})
 
 var port = process.env.PORT || 3000
 
@@ -39,13 +37,14 @@ app.get('/', function (req, res) {
 
 var router = express.Router()
 
+router.use(auth.ensureAuthenticated)
+
 router.use('/videos', require('./routes/videos.js'))
 router.use('/users', require('./routes/users.js'))
 
 router.get('/', function (req, res) {
   res.json({ message: 'API' })
 })
-
 
 app.use('/auth', require('./routes/auth.js'))
 app.use('/api', router)
