+++ resources/passport.js
@@ -4,25 +4,25 @@ var User = require('../models/user')
 
 var passport = require('passport')
 
-  passport.serializeUser(function (user, done) {
-    done(null, user.userId)
-  })
+passport.serializeUser(function (user, done) {
+  done(null, user.userId)
+})
 
-  passport.deserializeUser(function (id, done) {
-    User.find({ 'userId': id }, function (err, user) {
-      done(err, user)
-    })
+passport.deserializeUser(function (id, done) {
+  User.find({ 'userId': id }, function (err, user) {
+    done(err, user)
   })
+})
 
-  passport.use(new GoogleStrategy({
-    clientID: process.env.GOOGLE_AUTH_CLIENT_ID,
-    clientSecret: process.env.GOOGLE_AUTH_SECRET_KEY,
-    callbackURL: process.env.GOOGLE_AUTH_CALLBACK_URL
-  },
+passport.use(new GoogleStrategy({
+  clientID: process.env.GOOGLE_AUTH_CLIENT_ID,
+  clientSecret: process.env.GOOGLE_AUTH_SECRET_KEY,
+  callbackURL: process.env.GOOGLE_AUTH_CALLBACK_URL
+},
   function (token, refreshToken, profile, done) {
     console.log(profile)
     process.nextTick(function () {
-      User.findOne({ 'google.id' : profile.id }, function (err, user) {
+      User.findOne({ 'google.id': profile.id }, function (err, user) {
         if (err) {
           return done(err)
         }
@@ -40,7 +40,7 @@ var passport = require('passport')
 
           newUser.save(function (err, savedUser) {
             if (err) {
-              throw err;
+              throw err
             }
             return done(null, savedUser)
           })
